requirements.txt
absl-py==2.1.0
alembic==1.13.1
colorlog==6.8.2
contextlib2==21.6.0
contourpy==1.2.0
cycler==0.12.1
filelock==3.9.0
fonttools==4.49.0
fsspec==2023.4.0
greenlet==3.0.3
Jinja2==3.1.2
kiwisolver==1.4.5
Mako==1.3.2
MarkupSafe==2.1.3
matplotlib==3.8.3
ml_collections==0.1.1
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.19.3
nvidia-nvtx-cu11==11.8.86
optuna==3.5.0
packaging==23.2
pillow==10.2.0
pyparsing==3.1.1
python-dateutil==2.8.2
PyYAML==6.0.1
setproctitle==1.3.3
six==1.16.0
SQLAlchemy==2.0.27
sympy==1.12
torch==2.2.1+cu118
torchaudio==2.2.1+cu118
torchvision==0.17.1+cu118
tqdm==4.66.2
triton==2.2.0
typing_extensions==4.8.0
